Share mock Vercel response harness across push tests
